Remove thredds-test dependence from TestGribSubset. * Three tests that used to depend on TestDir.threddsTestServer now depend on the embedded TDS. The datasets used in the tests have been changed to ones available on the embedded TDS. * In order to make that work, those tests had to be moved from :cdm-test to :it. I created a new test class in :it named TestGridSubsetThredds. * All tests that require a TDS--whether embedded or remote--have been moved to the new class. * Converted log output from STDOUT to slf4j.
